refactor : 추천 게시물 server actions
